New content for section: WIS2 Architecture. (#61)
* New content for section: WIS2 Architecture.

Written to provide a high-level overview of things important to WIS2: roles and components.

References to the following resources are removed:
* sections/data-and-metadata-flows.adoc
* sections/real-time-data-exchange.adoc
* sections/discovery-metadata.adoc

I think these are no longer needed given the revamp of the Guide.

* update WIS2 Architecture section following review by efucile

Responded to comments from @efucile review.

---

> "WIS2 Nodes, GISCs and Global Services are types of WIS Centre" ...

Updated to better reflect what the manual says. In (new) line 11 I say that a WIS Centre may operate one or more global services. Still true - but a little bit looser terminology than the manual. CA-ECCC isn't a GISC :)

> "Global Discovery Catalogue: enables users (humans and software agents)" ...

Removed the caveat about "humans and software agents" to avoid the confusion.

> 3.7.5.6 A Global Cache shall retain a copy of core data for a duration...

Amended the text here to use what's said in the manual.